+++ hal/network/ncp/cellular/ncp.h
@@ -1,6 +1,5 @@
 /*
- ******************************************************************************
- *  Copyright (c) 2015 Particle Industries, Inc.  All rights reserved.
+ * Copyright (c) 2019 Particle Industries, Inc.  All rights reserved.
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -14,8 +13,14 @@
  *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, see <http://www.gnu.org/licenses/>.
- ******************************************************************************
  */
 
-// TODO: Move this header to a common directory
-#include "../../electron/modem/enums_hal.h"
+#pragma once
+
+#include "cellular_network_manager.h"
+
+namespace particle {
+
+CellularNetworkManager* cellularNetworkManager();
+
+} // particle