+++ hal/network/ncp/cellular/network_config_db.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Particle Industries, Inc.  All rights reserved.
+ * Copyright (c) 2019 Particle Industries, Inc.  All rights reserved.
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -15,15 +15,14 @@
  * License along with this library; if not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef HAL_BORON_NETWORK_NCP_H
-#define HAL_BORON_NETWORK_NCP_H
+#pragma once
 
-#include "cellular_network_manager.h"
+#include <cstddef>
 
 namespace particle {
 
-CellularNetworkManager* cellularNetworkManager();
+class CellularNetworkConfig;
+
+CellularNetworkConfig networkConfigForImsi(const char* imsi, size_t size);
 
 } // particle
-
-#endif // HAL_BORON_NETWORK_NCP_H