+++ system/src/control/wifi_new.cpp
@@ -22,9 +22,9 @@
 
 #include "common.h"
 
-#include "network/ncp.h"
-#include "wifi_network_manager.h"
-#include "wifi_ncp_client.h"
+#include "network/ncp/wifi/ncp.h"
+#include "network/ncp/wifi/wifi_network_manager.h"
+#include "network/ncp/wifi/wifi_ncp_client.h"
 
 #include "scope_guard.h"
 #include "check.h"