+++ system/src/system_network_manager.cpp
@@ -29,8 +29,8 @@ LOG_SOURCE_CATEGORY("system.nm")
 #include "enumclass.h"
 #include "system_commands.h"
 #if HAL_PLATFORM_WIFI && HAL_PLATFORM_NCP
-#include "network/ncp.h"
-#include "wifi_network_manager.h"
+#include "network/ncp/wifi/ncp.h"
+#include "network/ncp/wifi/wifi_network_manager.h"
 #endif // HAL_PLATFORM_WIFI && HAL_PLATFORM_NCP
 #if HAL_PLATFORM_NCP && HAL_PLATFORM_CELLULAR
 #include "cellular_hal.h"